Move VerifyingRpcClient tests into their own module
